+++ examples/shared_samples/firmware_driver.rb
@@ -16,14 +16,12 @@ require_relative '../_client' # Gives access to @client
 # The OneView appliance must have a valid spp and hotfix
 #
 # Supported APIs:
-# - 200, 300, 500
+# - 200, 300, 500, 600, 800, 1000, 1200, 1600
+# Supported API variants
+# C7000, Synergy
 
-# Resources that can be created according to parameters:
-# api_version = 200 & variant = any to OneviewSDK::API200::FirmwareDriver
-# api_version = 300 & variant = C7000 to OneviewSDK::API300::C7000::FirmwareDriver
-# api_version = 300 & variant = Synergy to OneviewSDK::API300::Synergy::FirmwareDriver
-# api_version = 500 & variant = C7000 to OneviewSDK::API500::C7000::FirmwareDriver
-# api_version = 500 & variant = Synergy to OneviewSDK::API500::Synergy::FirmwareDriver
+# for example, if api_version = 800 & variant = C7000 then, resource that can be created will be in form
+# OneviewSDK::API800::C7000::FirmwareDriver
 
 # Resource Class used in this sample
 fw_driver_class = OneviewSDK.resource_named('FirmwareDriver', @client.api_version)