+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This release extends support of the SDK to OneView REST API version 1600 (OneVie
 #### Features supported
 - Connection Template
 
+#### Bug fixes & Enhancements
+- [#432](https://github.com/HewlettPackard/oneview-sdk-ruby/issues/432)Snapshot deletion will fail with mis match etags error
+
 ## v5.12.0
 
 #### Notes