+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v5.12.1 (unreleased)
+## v5.13.0
 
 #### Notes
 This release extends support to OneView Rest API version 1600 (OneView v5.20) for the hardware variants C7000 and Synergy to the following features: